requirements.txt
Flask==0.10.1

# Application server for both development and production.
gunicorn==19.4.5

# Forms.
Flask-WTF==0.9.5
WTForms-Components==0.9.7
requests==2.18.4
Flask-OAuth==0.12